requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.30
mysql-connector-python==8.4.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.6

